% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_comb.R
\name{all_output}
\alias{all_output}
\title{GLOBIOM output combinations}
\usage{
all_output(df_gl)
}
\arguments{
\item{df_gl}{dataframe with globiom output.}
}
\value{
A data.frame with existing VAR_ID, VAR_UNIT, ITEM_AG combinations in df_gl
}
\description{
Creates data frame with all existing combinations of VAR_ID, VAR_UNIT and ITEM_AG in GLOBIOM output file.
}
\examples{
all_output(globiom)

}
