% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_nuts_pdf.R
\docType{data}
\name{eu_nuts_pdf}
\alias{eu_nuts_pdf}
\title{Potentially Disappeared Fraction of global species at EU NUTS level.}
\format{A data frame with 1476 rows and 3 variables:
\describe{
  \item{NUTS2}{NUTS2 ID}
  \item{scenario}{three scenario runs}
  \item{year}{year}
  \item{value}{PDF/ha}

}}
\source{
Di Fulvio et al. (2019), Spatially explicit LCA analysis of biodiversity
losses due to different bioenergy policies in the European Union, Science of the
Total Environment, 651, 1505-1516
}
\usage{
eu_nuts_pdf
}
\description{
A dataset containing a measure of Potentially Disappeared Fraction of
global species per ha at EU NUTS level from Di Fulvio (2019).
}
\keyword{datasets}
