% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globiomtools.R
\name{expand_output}
\alias{expand_output}
\title{Combinations of all output dimensions}
\usage{
expand_output(df_gl)
}
\arguments{
\item{A}{dataframe with globiom output}
}
\description{
Creates a dataframe with all variable, item and unit combinations in the data
}
\details{
This is additional  information
}
