% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globiomvis-package.R
\docType{package}
\name{globiomvis-package}
\alias{globiomvis}
\alias{globiomvis-package}
\title{globiomvis: Visualize GLOBIOM data with R}
\description{
Functions and examples for visualizing GLOBIOM data with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iiasa.github.io/globiomvis}
  \item \url{https://github.com/iiasa/globiomvis}
  \item Report bugs at \url{https://github.com/iiasa/globiomvis/issues}
}

}
\author{
\strong{Maintainer}: Michiel van Dijk \email{vandijkm@iiasa.ac.at} (\href{https://orcid.org/0000-0002-5207-7304}{ORCID})

Authors:
\itemize{
  \item Albert Brouwer
}

Other contributors:
\itemize{
  \item Stefan Frank [contributor]
  \item Fulvio Di Fulvio [contributor]
}

}
\keyword{internal}
