% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globiomvis.R
\docType{package}
\name{globiomvis}
\alias{globiomvis}
\alias{globiomvis-package}
\title{globiomvis: A package to facilitate GLOBIOM analysis.}
\description{
The globiomvis package provides a number of functions and examples to support the analysis with GLOBIOM,
including creating plots that compare historical trends with GLOBIOM projections and making maps.
}
\details{
At the moment the package describes the following tasks:
\itemize{
 \item Generating GLOBIOM scenario plots that compare historical trends with GLOBIOM projections.
 \item Generating EU NUTS maps.
 \item Generating a global map with the 30 GLOBIOM regions.
 \item Generating a global map using LUIDs as the unit of observation.
}
}
