% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_output_comb}
\alias{main_output_comb}
\title{Main VAR_ID, VAR_UNIT and ITEM_AG combinations.}
\format{A data frame with 272 rows and 3 variables:
\describe{
  \item{VAR_ID}{GLOBIOM variable code}
  \item{VAR_UNIT}{GLOBIOM variable unit}
  \item{ITEM_AG}{GLOBIOM item code}
}}
\source{
GLOBIOM
}
\usage{
main_output_comb
}
\description{
Main combinations of VAR_ID, VAR_UNIT and ITEM_AG to inspect GLOBIOM output.
}
\examples{
main_output_comb
}
\keyword{datasets}
