% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region30_country.R
\docType{data}
\name{region30_country}
\alias{region30_country}
\title{GLOBIOM 30-region to country mapping.}
\format{Data frame with 182 rows and 5 variables:
\describe{
  \item{ANYREGION}{GLOBIOM region label}
  \item{ALLCOUNTRY}{GLOBIOM country label}
  \item{iso3c}{ISO 3-letter country code}
  \item{country.name}{Country name}
  \item{GLOBIOM}{'y' for GLOBIOM-originated}
}}
\source{
GLOBIOM
}
\usage{
region30_country
}
\description{
GLOBIOM 30-region to country mapping.
}
\keyword{datasets}
