% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gams.R
\name{search_gams}
\alias{search_gams}
\title{GLOBIOMtools_env specific environmental variables and paths}
\usage{
search_gams(path = NULL, cmd = ifelse(Sys.info()["sysname"] ==
  "Windows", "gams.exe", "gams_cmd"), root = NULL)
}
\arguments{
\item{search_path}{Character. Force a search in a specified directory. This directory should contain the gdalinfo(.exe) executable. If a valid GDAL install is found in this path, this will force gdalUtils to use this installation. Remember to set rescan=TRUE if you have already set an install.}

\item{rescan}{Logical. Force a rescan if neccessary (e.g. if you updated your GDAL install).}

\item{ignore.full_scan}{Logical. If FALSE, perform a brute-force scan if other installs are not found. Default is TRUE.}

\item{verbose}{Logical. Enable verbose execution? Default is FALSE.}
}
\value{
Location of folder where most recent version of gams.exe is stored.
}
\description{
Creates scenario plots by region for selected items
}
\details{
Sets the environmental, package specific parameters and settings.
}
\examples{
search_gams()
Search GAMS path

Searches for the path where gams.exe is stored

This is additional  information

}
