% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{This function allows you to plot historical trends and globiom results}
\usage{
trend_plot(var_sel, item_sel, unit_sel, reg_sel, df_gl = globiom,
  df_hs = hist, display = T, output = T, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to plot historical trends and globiom results
}
\examples{
cat_function()
}
\keyword{cats}
