% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot_all}
\alias{trend_plot_all}
\title{Creates a PDF with plots for selected GLOBIOM output dimensions}
\usage{
trend_plot_all(df_gl, df_hs, path = NULL, file_name = NULL, comb = NULL)
}
\arguments{
\item{df_gl}{Data frame with globiom output.}

\item{df_hs}{Data frame with historical data aggregated to GLOBIOM nomenclature.}

\item{path}{Path where the PDF file will be saved. Default is the working
directory.}

\item{file_name}{Name of the PDF file that is generated. Default is
\code{globiom_trend_plots_YYYY-MM-DD.pdf}.}

\item{comb}{Output combinations used for plotting. Default is
\code{all_output_comb(df_gl)} as present in main_output_comb.}
}
\value{
None but a PDF file is saved in the working directory or a specified
  location.
}
\description{
Creates a PDF file with plots for selected GLOBIOM VAR_ID, ITEM_AG and
VAR_UNIT combinations.
}
\details{
Plots are produced for a standard set of key VAR_ID, ITEM_AG and VAR_UNIT
combinations stored in \code{output_comb_base} unless specified otherwise.
Plots are only produced for output combinations that exist in the data.
}
\examples{
\dontrun{
trend_plot_all(df_gl = globiom,
               df_hs = hist,
               path = "c:/temp",
               file_name = "globiom_output")
}

}
