% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{Create plots for GLOBIOM scenarios}
\usage{
trend_plot(var_sel, item_sel, unit_sel, reg_sel, df_gl = globiom,
  df_hs = hist, display = T, output = T, ...)
}
\arguments{
\item{df_gl}{dataframe with globiom output}

\item{df_hs}{dataframe with historical data}

\item{var}{Variable}

\item{itme}{Item}

\item{unit}{Unit}

\item{reg}{Region}
}
\description{
Creates scenario plots by region for selected items
}
\details{
This is additional  information
}
