% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{Create plots for GLOBIOM scenarios}
\usage{
trend_plot(var, item, unit, reg, df_gl, df_hs, display = T, output = T, ...)
}
\arguments{
\item{var}{VAR_ID}

\item{item}{ITEM_AG}

\item{unit}{VAR_UNIT}

\item{reg}{REGION_AG}

\item{df_gl}{data frame with globiom output}

\item{df_hs}{data frame with historical data aggregated to GLOBIOM
nomenclature}
}
\value{
A ggplot object
}
\description{
Creates scenario plots by region for selected items.
}
\details{
Note that the function uses standard GLOBIOM variable (VAR_ID), item
(ITEM_AG), unit (VAR_UNIT) and region (REGION_AG) nomenclature as input.
VAR_ID, ITEM_AG and VAR_UNIT have to be in upper case!
}
\examples{
\dontrun{
trend_plot(var = "LAND", item = "1000 HA", reg = "World", df_gl = globiom, df_hs = hist)
}

}
