% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot_all}
\alias{trend_plot_all}
\title{Creates pdf with plots for selected GLOBIOM output dimensions}
\usage{
trend_plot_all(df_gl = NULL, df_hs = NULL, path = NULL,
  file_name = NULL, output_sel = NULL)
}
\arguments{
\item{df_gl}{dataframe with globiom output}

\item{df_hs}{data frame with historical data aggregated to GLOBIOM nomenclature}

\item{path}{path where the pdf file will be saved. Default is the working directory.}

\item{file_name}{name of the pdf file that is generated. Default is
\code{globiom_trend_plots_YYYY-MM-DD.pdf}.}
}
\value{
A dataframe
}
\description{
Creates a pdf file with plots for selected GLOBIOM VAR_ID, ITEM_AG and VAR_UNIT combinations.
}
\details{
Plots are produced for a standard set of key VAR_ID, ITEM_AG and VAR_UNIT combinations
stored in \code{output_comb_base} unless specified otherwise
}
\examples{
trend_plot_all(df_gl = globiom, df_hs = hist, path = "c:/temp", file_name = "globiom_output",
output_se = df)

}
